Moved bot owner commands to Bot class
